requirements.txt
langchain
langchain-community
langgraph
requests
sentence-transformers
chromadb
googlesearch-python
beautifulsoup4